% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/counterfactual_f.R
\name{estim_counterfactual}
\alias{estim_counterfactual}
\title{Calculate counterfactual}
\usage{
estim_counterfactual(data,
                           model,
                           period,
                           alpha)
}
\arguments{
\item{data_input}{The data containing the variables in the model,
and a named column date with date time data.
If no dates column is provided, then the first row is assumed to be a date.}

\item{chosen_model}{A model for which scenarios are simulated}

\item{chosen_period}{The period when the  counterfactual scenario is simulated.
The period covered by data_input should cover the chosen_period.}

\item{alpha}{Optionally, the lag effect between (0,1). If it is not specified, it is set to zero.}
}
\value{
Return counterfactual values
}
\description{
Simulate counterfactual scenarios.
}
\details{
Most scenarios are constructed with historical actuals
}
\examples{
## Calculate counterfactual scenarios under different scenarios
data("sales_data")
sales_model1 <- lm(total ~
wday +
  month +
  adj_med_spend,
data = sales_data \%>\%
  filter(date < as.Date("2018-07-01"))
)

start_campaign <- as.Date("2018-07-01")
end_campaign <- as.Date("2018-07-25")
chosen_period <- start_campaign + seq(0, end_campaign - start_campaign)

## Scenario with no Media Spend
sales_data \%>\%
  mutate(adj_med_spend = min(adj_med_spend)) \%>\%
  estim_counterfactual(sales_model1, chosen_period) \%>\%
  sum()

## Sales under the scenario that we haven't done any marketing campaign
sales_data \%>\%
  estim_counterfactual(sales_model1, chosen_period) \%>\%
  sum()
}
