% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/counterfactual_f_part2.R
\name{marketing_effects}
\alias{marketing_effects}
\title{Estimate effects of different variables}
\usage{
marketing_effects(data_input, scenarios)
}
\description{
Estimate effects of different variables
}
