% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulations_f.R
\name{month_trend}
\alias{month_trend}
\title{Month trend}
\usage{
month_trend(month, baseline)
}
\arguments{
\item{baseline}{A vector of values per day}

\item{weekday}{Day of the week as Date, word in English or number:}
}
\description{
Calculate counterfactual under different scenarios.
}
\examples{
## Example
month_trend(month='October',
           baseline=50:62)
month_trend(month=10,
           baseline=50:62)
month_trend(month=as.Date('2019-10-25'),
           baseline=50:62)
}
