% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliary_f_part2.R
\name{pivot}
\alias{pivot}
\title{Pivot Table}
\usage{
pivot(data_input, columns, groups)
}
\arguments{
\item{data_input}{Data}

\item{columns}{Columns representing different values}

\item{groups}{Name of the groups}
}
\description{
Data transformation
}
\examples{
## Calculate counterfactual scenarios under different scenarios

data("sales_data")
sales_data \%>\%
  pivot(c("cs1", "cs2"), c("cs1", "cs2")) \%>\%
  filter(date \%in\% c(baseline_period, campaign_period)) \%>\%
  ggplot() +
  geom_line(aes(x = date, y = total, col = group)) +
  geom_vline(xintercept = min(campaign_period))
}
