% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/counterfactual_f_part2.R
\name{variable_mix}
\alias{variable_mix}
\title{Calculate Mix}
\usage{
variable_mix(data_input, chosen_period = NA, target_var = NA)
}
\arguments{
\item{data_input}{A tibble table with a column date
having time information and a column  target_var grouping the entries.
If no target_var is provided, then specify the name of target_var.}

\item{chosen_period}{Period where the mix is calculated}

\item{target_var}{Variable having group information}
}
\value{
variable mix returns a summary a pivot table with the mix total
}
\description{
Calculate Mix


}
\examples{

data("Sales")
period_mix<-as.Date('2019-01-03')+seq(0,13)
Sales\%>\%
 filter(year(date)==2019 & week(date) \%in\% 1:5)\%>\%
 variable_mix(chosen_period=period_mix,target_var='SALES_CHANNEL')

}
