% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulations_f.R
\name{week_trend}
\alias{week_trend}
\title{Simulate weekdays}
\usage{
week_trend(weekday, baseline)
}
\arguments{
\item{weekday}{Day of the week as Date, word in English or number:}

\item{baseline}{A vector of values per day}
}
\description{
Calculate counterfactual under different scenarios.
}
